Export policy_rule and policy_rule_func decorators
Previously, `policy_rule_func` was an internal-only decorator that was
parameterized by the underlying (transformed) monad.

In order to implement any low-level policy rule functions external to the
library, the implementer would have also been forced to import
pymonad.List.

Instead, to avoid that external dependency, as well as requiring an unnecessary
parameter:

- Rename `policy_rule_func` to `policy_rule_funcM` and define a new decorator
  with that name, behaving in the same way, except assuming StateT(List)

- Define `policy_rule` decorator for symmetry, in lieu of the awkwardly
  named `policyM`